Toolbox error dialog now has version number in title.
Side effect: "java -jar iText.jar" is enough to get the version number.

If the toolbox jar is present, the GUI can display the versions numbers
using "Help->Versions".

If the toolbox is not there, the iText version is in the dialog title.

This may help with customer support queries...


git-svn-id: http://svn.code.sf.net/p/itext/code/trunk@2966 3a709846-db2a-4fbd-be0b-eb932eee6479