pattern for test cases was too restrictive, some tests were not run
Conversion of TestPdfCopyAndStamp to use JUnit 4 annotations instead of
inheriting from TestCase

Two tests in TestPdfCopyAndStamp are tagged with @Ignore
as they result in failure or error. It looks like they only work
on Windows anyway and with a GUI so not suitable for automatic
testing. Maybe I'll move them to the test.swing part later on


git-svn-id: http://svn.code.sf.net/p/itext/code/trunk@3663 3a709846-db2a-4fbd-be0b-eb932eee6479